Move the Uart mock into the common bsw mock folder
