+++ executables/unitTest/bsp/CMakeLists.txt
@@ -1,4 +1,3 @@
 add_subdirectory(bspConfiguration)
 add_subdirectory(bspIo)
 add_subdirectory(bspMcu)
-add_subdirectory(bspUart)
