--- executables/unitTest/bsp/bspUart/CMakeLists.txt
@@ -1,5 +1,0 @@
-add_library(utbspUart src/bsp/Uart.cpp)
-
-target_include_directories(utbspUart PUBLIC include)
-
-target_link_libraries(utbspUart PRIVATE bsp bspConfiguration)
