--- executables/unitTest/bsp/bspUart/doc/index.rst
@@ -1,9 +1,0 @@
-bspUart
-=======
-
-Overview
---------
-
-The ``bspUart`` module consist of empty implementation of the
-functions defined in ``bspUart`` module, intended to be used only for
-building unit test cases.
