--- executables/unitTest/bsp/bspUart/module.spec
@@ -1,2 +1,0 @@
-maturity: raw
-oss: true
