+++ platforms/s32k1xx/bsp/bspUart/test/CMakeLists.txt
@@ -2,6 +2,6 @@ add_executable(bspUartTest src/IncludeTest.cpp)
 
 target_include_directories(bspUartTest PRIVATE mock/include)
 
-target_link_libraries(bspUartTest PRIVATE utbspUart utbspIo gmock gtest_main)
+target_link_libraries(bspUartTest PRIVATE bspMock utbspIo gmock gtest_main)
 
 gtest_discover_tests(bspUartTest PROPERTIES LABELS "bspUartTest")
