+++ platforms/s32k1xx/bsp/bspUart/test/src/IncludeTest.cpp
@@ -1,4 +1,3 @@
-#include "UartMock.h"
 #include "bsp/Uart.h"
 
 #include <gtest/gtest.h>
@@ -7,7 +6,7 @@ namespace
 {
 TEST(BspTest, IncludeCheck)
 {
-    bsp::UartMock uartMock;
+    bsp::Uart uartMock;
     uartMock.write(::etl::span<uint8_t const>{});
     uartMock.read(::etl::span<uint8_t>{});
 }
